Use cloud-prepare for AWS provisioning
This is a first step towards delegating all cloud provisioning to
cloud-prepare. For now, we only set up the cloud-prepare client; all
the actual AWS configuration still relies on the addon-specific code.

The Submariner dependencies need to be updated to the current devel
versions (with the associated transitive dependencies) to work with
the K8s 1.19+ APIs.

Signed-off-by: Stephen Kitt <skitt@redhat.com>