Add UT hook to submarinerConfigController.sync
The sync function is invoked in a goroutine so in order for gomock
to gracefully report unexpected calls, we need to invoke GinkgoRecover
as a deferred function. Add a hook for the UTs to do that. To simplify
the NewSubmarinerConfigController function, the params are now passed
via a SubmarinerConfigControllerInput struct.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>