Remove use of deprecated/removed Gingko 2.0 APIs
Gingko emits warnings about upcoming deprecated or removed
functionality in Gingko 2.0 being used in the test/integration suite:

- Removal of custom reporters as outlined here:
https://github.com/onsi/ginkgo/blob/ver2/docs/MIGRATING_TO_V2.md#removed-custom-reporters

This is due to the use of the printer.NewlineReporter in
TestIntegration. It doesn't seem necessary to use this.

- Removal of async testing as outlined here:
https://github.com/onsi/ginkgo/blob/ver2/docs/MIGRATING_TO_V2.md#removed-async-testing

This is due to the use of the Done channel in the BeforeSuite
function to run it async with a minute time out. Perhaps this was added
b/c testEnv.Start ran too long or hung at some point previously. We can
address this later if it ever occurs.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>