Create ClusterRole to allow brokers.submariner.io object
This PR installs a ClusterRole which when associated with a
roleBinding will allow the respective user to create/update
the brokers.submariner.io object. The cluster admin is
expected to assign the roleBinding to the ManagedClusterSetAdmin
who is interested in installing Submariner addon on the
managedClusters in the clusterSet.

Related to: https://github.com/submariner-io/enhancements/issues/51
Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>