Create a ClusterRole to allow brokers.submariner.io object
Starting with ACM 2.5, a user who is installing Submariner
Addon on the ManagedCluster should have permissions to
create/update/read brokers.submariner.io object in the
ManagedClusterSet broker namespace.
This object needs to be created whether Globalnet is
enabled or not.

A clusterAdmin has necessary privileges to create this
object, but for a non-ClusterAdmin, an explicit role has
to be assigned. This PR creates a ClusterRole as part of
Submariner addon initialization. Assigning this ClusterRole
to the respective user has to be done manually.

Fixes: https://github.com/stolostron/submariner-addon/issues/321
Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>