Remove apiextensions/v1beta1 reference
Re: https://github.com/stolostron/backlog/issues/23199, the only
deprecated K8s API listed in
https://kubernetes.io/docs/reference/using-api/deprecation-guide/#v1-22
that is used/referenced in the code is
"k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1". This API
is only referenced to add it to the scheme and is not actually used so
remove it.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>