Fix intermittent deployment_controller test failure
/pkg/spoke/submarineragent/deployment_controller_test.go:249

      Expected
          <v1.ConditionStatus>: False
      to equal
          <v1.ConditionStatus>: True

Before we get the globalnet deployment status, we get the Submariner
resource to check if globalnet is enabled. We used the queue key to
extract the namespace and name however that's only valid depending on
which resource triggered the update. Instead, use the known installation
namespace to list the Submariner resources and pick the first one (there
shoild only be one).

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>