Move default provider handling to the spoke
...specifically the case where there's no cloud provider
implementation for a platform. We still want to set the status
condition indicating the environment was prepared. Previously
this was done on the hub but it should be done on the spoke as
only AWS is handled on the hub (the intention is to move this
to the spoke as well).

The cloud.Providerfactory was modified to also return a boolean
indicating whether or not a Provider exists.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>