Create the addon SA before the deployment
The agent uses a ManifestWork to deploy the agent on the hub; this
ManifestWork is charged with creating the roles, role bindings,
service account, and deployment for the addon. In some cases, the
deployment fails because the service account hasn't been created by
the time the pod replicaset is instantiated.

This reorders the ManifestWork entries so that the SA comes first,
followed by the roles and bindings, and the deployment comes last.
This should ensure that the SA is created before any resource which
needs it.

Signed-off-by: Stephen Kitt <skitt@redhat.com>