Restart addon pod on managed cluster after restore
After a backup+restore, submariner-addon pods on managed cluster are
still trying to reach old hub. The pods need to be restarted to pick
up new hub config.

Add an env variable to submariner-addon deployment with hub's API
server. When new hub is restored, it will result in forced restart
of the pods on managed cluster.

Fixes #584

Signed-off-by: Vishal Thapar <5137689+vthapar@users.noreply.github.com>