Add backup label to submariner-broker resource
Submariner broker resource is not backed up if it is missing
backup label. This can break submariner in case of backup+restore.
Broker object is only created when submariner is installed on a
managed cluster. Check for label when deploying submariner-agent and
add if not already present.

Related to stolostron/backlog#22005

Signed-off-by: Vishal Thapar <5137689+vthapar@users.noreply.github.com>