Extract submver from dependencies
Instead of updating submver manually, use whatever is specified as the
Submariner dependency. This avoids having to update a variable every
time Submariner is released, which means we can rely on dependabot to
take care of everything for us.

While we're at it, drop the Submariner CR update; it no longer
specifies the version to use.

Signed-off-by: Stephen Kitt <skitt@redhat.com>