Remove the submariner-agent-scc
This is associated with all the submariner ServiceAccounts to
enable the pods to have privileged access. However this is no
longer needed since the ServiceAccounts are associated with
the built-in privileged SCC via ClusterRoles, which are included
with the submariner bundle.

Also the submariner-agent-scc was seen to be problematic in one
installation as it is also associated with several system groups
and thus can interfere with admission of non-submariner pods since
it has priority 1.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>