Ensure ManagedCluster finalizer removed if add-on does not exist
The submarinerAgentController assumes the submariner ManagedClusterAddOn
exists when cleaning up on ManagedCluster deletion. However, if it
doesn't, cleanup never completes and the finalizer is not removed from
the ManagedCluster. This shouldn't normally happen since the controller
aso adds a finalizer to the ManagedClusterAddOn but this scenario was
observed in a real installation
(see https://issues.redhat.com/browse/ACM-5381).

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>
(cherry picked from commit 4685987c45ba1364a9687d57bb0740274314c6cc)