Ensure the agent has a writable /tmp
When run with read-only root, the agent fails because it expects to be
able to write to /tmp. This adds a writable /tmp.

The addon's specific use of a writable /tmp is tied to certificate
provisioning, and the correct long-term fix is to manage its
certificates externally; but until that's done, providing a writable
/tmp has no adverse effects and allows read-only root deployments.

Signed-off-by: Stephen Kitt <skitt@redhat.com>