Fix spoke agent test case
The test attempts to update the Submariner status and verify that the
ManagedClusterAddOn status is updated. But spoke controllers were failing
to access the API server due to certificate issue. The was due to the
"fake" hub kube config file used to create the rest config. Using the
test environment rest config fixes the issue so the AgentOptions was
modified to allow the rest config to be explicitly set.

Despite the API server issue, the test was still passing. This was due
to incorrect usage of meta.IsStatusConditionTrue.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>