Fix concurrent map read/write panic in resourceapply module
In resource.ApplyManifests, we call resourceapply.ApplyDirectly and pass
in a global ResourceCache. However we have 2 controllers that use this
function which can result in concurrent access to the ResourceCache as
the controllers run on different threads. This was observed in a recent
integration test run. To alleviate this, modify each controller to have
its own ResourceCache instance and pass to resource.ApplyManifests.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>