On RHOSP, disable verification if requested
The current cloud setup on RHOSP ignores the "verify" setting in
clouds.yaml. On setups with self-signed certificates, this means that
the addon can't connect to the RHOSP API endpoint.

When "verify" is set to false, this sets up a custom TLS configuration
with disabled verification.

Ideally a complete RHOSP implementation would also consider the
certificate information which can be specified in clouds.yaml; see
PrepareTLSConfig in
https://github.com/gophercloud/utils/blob/master/internal/util.go for
details. It would be necessary to determine a way for the files to be
provided (in the same way as clouds.yaml).

github.com/gophercloud/utils also supports this, but only through
NewServiceClient, not the general-purpose AuthenticatedClient
function. Using NewServiceClient would require reworking
cloud-prepare.

Signed-off-by: Stephen Kitt <skitt@redhat.com>