Revert to non-Konflux builds by default
Konflux builds seem to rely on Tekton directives rather than the
Makefile. Brew-based container builds are now failing, either because
of the switch to Dockerfile.konflux by default, or the use of
-mod=mod. To re-enable Brew-based builds, control the latter from
Dockerfile.konflux, and default to Dockerfile from the Makefile. The
Tekton build configuration explicitly refers to Dockerfile.konflux so
this shouldn't affect Konflux builds.

Signed-off-by: Stephen Kitt <skitt@redhat.com>