Delete cluster-specific broker resources on add-on removal
If the associated managed cluster doesn't have connectivity to the
hub then Submariner uninstall won't occur and its broker resources
(Endpoint, Cluster etc) will remain which can cause issues.
Modified the add-on agent on the hub to also clean up the broker
resources after the ManifestWorks are deleted as a backup. Also,
the ManifestWork controllers are deployed on the managed cluster
and they add finalizers so the ManifestWorks may not get deleted.
To handle this case, a 2 min timeout was added to force broker
cleanup if the ManifestWorks aren't deleted.

See https://issues.redhat.com/browse/ACM-11656

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>