Orphan installation namespace for the addon deploy ManifestWorks
The addon controller is deployed in the same namespace (defaults to
"submariner-operator") as the Submariner components on the spoke cluster.
The ManifestWorks deploys the Namespace resource as well to ensure it's
created when deploying the other resources however we don't want the
ManifestWorks to delete the Namespace on uninstall to avoid a race
condition where the Submariner operator pod is deleted before it is
able to run cleanup and remove its finalizer from the Submariner
resource.

Set the addon.open-cluster-management.io/deletion-orphan annotation
on the Namespace resource so the addon framework configures the
ManifestWorks to orphan it.

See https://issues.redhat.com/browse/ACM-15538 for more details.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>