Add operator Namespace to the Subscription manifest work
...so the namespace is deleted on clean up. The namesapce used to get
deleted via the agent add-on manifest work but that was removed b/c it
interfered with submariner cleanup on the managed clusters. It was
thought that deleting the Subscription would cause the installed
operator components to be deleted but that's not the case - deleting
a Subscription only unsubscribes from future updates and leaves the
installation intact.

Related to https://issues.redhat.com/browse/ACM-16491

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>