Add support for configurable resource limits and requests
...for the submariner add-on agent deployment following the guidelines
described here:
https://github.com/stolostron/foundation-docs/blob/main/addon/addon-resource-requirements.md

This entails migrating to the AgentAddonFactory API which creates an
AgentAddon implementation from various configuration options. This
does all the work of creating the manifest resources from the
template YAML files and handling custom NodeSelector, Tolerations,
and ResourceRequirements configured in an AddOnDeploymentConfig.
We just need to place the proper template substitution code in
the deployment template YAML file.

The manifests directory under submarineraddonagent was split into
two sub directories, spoke and hub, to avoid the AgentAddonFactory
code from including the hub-specific RBAC resources in the agent
manifest.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>