Switch to Go 1.24 tool handling
Instead of specifying tool versions in the Makefile, this adds tool
dependencies in a dedicated module (to avoid intermingling
dependencies with the main project).

Instead of direct "go tool" invocations (which don't work well with a
separate module, unless an undocumented parameter is used) or local
cached binaries (which makes it hard to handle rebuilds when tool
versions change on systems with pre-existing versions of the tools),
this uses a gotool macro to retrieve the cached binary used by "go
tool". "go tool -n" is invoked twice because the first invocation
returns the path to the initial binary, which is a temporary binary
removed as soon as "go tool" completes; the second invocation returns
the path to the cached binary.

operator-sdk is more complicated to build, so the existing mechanism
is used; the only change there is that the tool dependency is
represented using "go tool" semantics in tools/operator-sdk/go.mod,
using another module because the old SDK version currently used is
incompatible with the other tools' dependencies.

Signed-off-by: Stephen Kitt <skitt@redhat.com>